+++ src/DependencyInjection/Configuration.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BlobConnectorCampusonlineDmsBundle\DependencyInjection;
 
+use Dbp\Relay\BlobLibrary\Api\BlobApi;
 use Dbp\Relay\CoreBundle\Authorization\AuthorizationConfigDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
@@ -19,6 +20,8 @@ class Configuration implements ConfigurationInterface
             ->append(AuthorizationConfigDefinition::create()
                 ->addRole(self::ROLE_USER, 'false', 'Returns true if the current user is authorized to use the API')
                 ->getNodeDefinition());
+
+        $treeBuilder->getRootNode()->append(BlobApi::getConfigNodeDefinition());
 
         return $treeBuilder;
     }