+++ src/DependencyInjection/DbpRelayBlobConnectorCampusonlineDmsExtension.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Dbp\Relay\BlobConnectorCampusonlineDmsBundle\DependencyInjection;
 
 use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Authorization\AuthorizationService;
+use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Service\DocumentService;
 use Dbp\Relay\CoreBundle\Extension\ExtensionTrait;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -26,6 +27,9 @@ class DbpRelayBlobConnectorCampusonlineDmsExtension extends ConfigurableExtensio
         $this->addResourceClassDirectory($container, __DIR__.'/../Entity');
 
         $definition = $container->getDefinition(AuthorizationService::class);
+        $definition->addMethodCall('setConfig', [$mergedConfig]);
+
+        $definition = $container->getDefinition(DocumentService::class);
         $definition->addMethodCall('setConfig', [$mergedConfig]);
     }
 }